fix: Change OffsetField type to short

+++ src/main/java/structures/image/OffsetField.java
@@ -7,5 +7,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class OffsetField {
     // Вообще-то тут 2 поля на 12 и 4 байт, но понятия не имею, как это адекватно сделать
-    long data;
+    short data;
 }
